Split trigger engine tests
